Add splash screen and update launch configurations
Added splash screen images for Android, iOS, and web platforms. Updated the launch background configurations for Android and iOS to use the new splash screen images. Also, configured the web index.html to include a splash screen that adapts to light and dark themes. Updated the `pubspec.yaml` to include the `flutter_native_splash` package and configured it accordingly.